% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_rain.R
\name{annual_rain}
\alias{annual_rain}
\title{Annual rainfall summaries}
\usage{
annual_rain(
  data,
  date_time,
  rain,
  year,
  total_rain = TRUE,
  n_rain = TRUE,
  rain_day = 0.85,
  station = NULL,
  na_rm = FALSE,
  na_prop = NULL,
  na_n = NULL,
  na_consec = NULL,
  na_n_non = NULL
)
}
\arguments{
\item{data}{}

\item{date_time}{}

\item{rain}{}

\item{year}{}

\item{total_rain}{}

\item{n_rain}{}

\item{rain_day}{}

\item{station}{}

\item{na_rm}{}

\item{na_prop}{}

\item{na_n}{}

\item{na_consec}{}

\item{na_n_non}{}
}
\value{
A data.frame with rainfall summaries for each year
}
\description{
Annual rainfall summaries
}
\examples{
#TODO
}
