% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_column_names.R
\name{assert_column_names}
\alias{assert_column_names}
\title{Check whether the columns are in the data}
\usage{
assert_column_names(data, columns)
}
\arguments{
\item{data}{Data frame to check the column is a part of}

\item{columns}{Name of columns to check}
}
\value{
A check whether the columns are in the data
}
\description{
Check whether the columns are in the data
}
