% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climatic_summary.R
\name{climatic_summary}
\alias{climatic_summary}
\title{Calculate summaries from climatic data}
\usage{
climatic_summary(
  data,
  date_time,
  station = NULL,
  elements = NULL,
  year = NULL,
  month = NULL,
  dekad = NULL,
  pentad = NULL,
  to = c("hourly", "daily", "pentad", "dekadal", "monthly", "annual-within-year",
    "annual", "longterm-monthly", "longterm-within-year", "station", "overall"),
  by = NULL,
  doy = NULL,
  doy_first = 1,
  doy_last = 366,
  summaries = c(n = "~dplyr::n()"),
  na_rm = FALSE,
  na_prop = NULL,
  na_n = NULL,
  na_consec = NULL,
  na_n_non = NULL,
  first_date = FALSE,
  n_dates = FALSE,
  last_date = FALSE,
  summaries_params = list(),
  names = "{.fn}_{.col}"
)
}
\arguments{
\item{data}{\code{data.frame} The data.frame to calculate from.}

\item{date_time}{\code{\link[base]{Date}} The name of the date column in \code{data}.}

\item{station}{\code{character(1)} The name of the station column in \code{data}, if the data are for multiple station.}

\item{elements}{\code{character} The name of the elements column in \code{data} to apply the function to.}

\item{year}{\code{character(1)} The name of the year column in \code{data}.}

\item{month}{\code{character(1)} The name of the month column in \code{data}.}

\item{dekad}{\code{character(1)} The name of the dekad column in \code{data}.}

\item{pentad}{\code{character(1)} The name of the pentad column in \code{data}.}

\item{to}{\code{character(1)} The date-time format to put the data into.}

\item{by}{\code{character} The name of columns in \code{data} to group the summary data by.}

\item{doy}{\code{character(1)} The name of the day of the year (1-366) column in \code{data}. If \code{NULL} it will be created using \code{lubridate::year(data[[doy]])}.}

\item{doy_first}{\code{integer(1)} The first day of the year.}

\item{doy_last}{\code{integer(1)} The last day of the year.}

\item{summaries}{\code{character} A named character vector of summary functions. The names are
used as the column names in the results. The values can be any function
name as a string or a function call as a formula. e.g. c(mean = "mean", st_dv = "sd", n = "~dplyr::n()")}

\item{na_rm}{\code{logical(1)} If \code{TRUE} all \code{na_} parameters are ignored and missing
values are removed. If \code{FALSE} missing values are not removed unless
any \code{na_} parameters are specified.}

\item{na_prop}{\code{integer(1)} Max proportion of missing values allowed}

\item{na_n}{\code{integer(1)} Max number of missing values allowed}

\item{na_consec}{\code{integer(1)} Max number of consecutive missing values allowed}

\item{na_n_non}{\code{integer(1)} Min number of non-missing values required}

\item{first_date}{\code{logical(1)} If \code{TRUE} the first instance of \code{date_time} when the value equals the summary value is included. Generally only used for extreme summaries i.e. first \code{date_time} when the maximum occurred.}

\item{n_dates}{\code{logical(1)} If \code{TRUE} the number of \code{date_time} points when the value equals the summary value is included. Generally only used for extreme summaries i.e. number of days in which the minimum occurred.}

\item{last_date}{\code{logical(1)} If \code{TRUE} the last instance of \code{date_time} when the value equals the summary value is included. Generally only used for extreme summaries i.e. last \code{date_time} when the maximum occurred.}

\item{summaries_params}{\code{list} Additional parameters to pass to \code{summaries}.
Must be a list of lists with the same names as \code{summaries}.}

\item{names}{Format of column names. Passed to \code{.names} in
\code{dplyr::across}}
}
\value{
A summary data frame for selected element(s) in climatic data.
}
\description{
\code{climatic_summary} returns a data table displaying
summary statistics for element(s) (and for each station) in a given time period.
}
\examples{
# Calculate frequencies for tmin/tmax for each year, month, and station.
# filter daily_niger data for the example
daily_niger_1 <- daily_niger \%>\% dplyr::filter(year > 1960)
climatic_summary(data = daily_niger_1, date_time = "date", station = "station_name",
                 elements = c("tmin", "tmax"), to = "monthly")
}
