% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crops_definitions.R
\name{crops_definitions}
\alias{crops_definitions}
\title{Crop Definitions}
\usage{
crops_definitions(
  data_name,
  year,
  station,
  rain,
  day,
  rain_totals,
  plant_days,
  plant_lengths,
  start_check = c("both", "yes", "no"),
  season_data_name,
  start_day,
  end_day,
  return_crops_table = TRUE,
  definition_props = TRUE,
  data_book = NULL
)
}
\arguments{
\item{data_name}{Name of the data frame containing the data.}

\item{year}{Name of the column representing the year.}

\item{station}{Name of the column for the station (optional).}

\item{rain}{Name of the column containing rainfall data.}

\item{day}{Name of the column containing day data.}

\item{rain_totals}{Column name for rain totals.}

\item{plant_days}{Column name for planting days.}

\item{plant_lengths}{Column name for planting lengths.}

\item{start_check}{Boolean indicating whether to check start day (default is TRUE).}

\item{season_data_name}{Name of the season data frame (optional).}

\item{start_day}{Column name for the start day.}

\item{end_day}{Column name for the end day.}

\item{return_crops_table}{Boolean indicating whether to return the full crops table (default is TRUE).}

\item{definition_props}{Boolean indicating whether to calculate properties (default is TRUE).}

\item{data_book}{A \code{DataBook} object to use. If \code{NULL}, a new one is created. Default \code{NULL}.}
}
\value{
TODO
}
\description{
Define crop conditions and create a new crop definition data frame.
}
\examples{
#TODO
}
