% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_rains.R
\name{end_rains}
\alias{end_rains}
\title{End of the rains}
\usage{
end_rains(
  data,
  date_time,
  station = NULL,
  year = NULL,
  rain = NULL,
  doy = NULL,
  s_start_doy = NULL,
  drop = TRUE,
  start_day = 1,
  end_day = 366,
  output = c("doy", "date", "status"),
  interval_length = 1,
  min_rainfall = 10,
  data_book = NULL
)
}
\arguments{
\item{data}{The data.frame to calculate from.}

\item{date_time}{\code{\link[base]{Date}} The name of the date column in \code{data}.}

\item{station}{\code{character(1)} The name of the station column in \code{data}, if the data are for multiple station.}

\item{year}{\code{character(1)} The name of the year column in \code{data}. If \code{NULL} it will be created using \code{lubridate::year(data[[date_time]])}.}

\item{rain}{\code{character(1)} The name of the rainfall column in \code{data} to apply the function to.}

\item{doy}{\code{character(1)} The name of the day of year column in \code{data} to apply the function to. If \code{NULL} it will be created using the \code{date_time} variable.}

\item{s_start_doy}{\code{numerical(1)} Default \code{NULL} (if \code{NULL}, \code{s_start_doy = 1}. The day of year to state is the first day of year.}

\item{drop}{\code{logical(1)} default \code{TRUE}. Whether to drop years where there are \code{NA} data for the rainfall.}

\item{start_day}{\code{numerical(1)} The first day to calculate from in the year (1-366).}

\item{end_day}{\code{numerical(1)} The last day to calculate to in the year (1-366).}

\item{output}{\code{character(1)} Whether to give the start of rains by day of year (doy), date, and/or status. Default all three selected.}

\item{interval_length}{\code{numerical(1)} Number of days for the minimum rainfall to fall in.}

\item{min_rainfall}{\code{numerical(1)} Minimum amount of rainfall to occur on the set of days defined in \code{interval_length}.}

\item{data_book}{The data book object where the data object is stored, default \code{NULL}.}
}
\value{
A data.frame with the day of year and/or date for the end of the rains for each year (and station).
}
\description{
Last occurrence from day q with at least z mm rainfall on a set of days.
}
\examples{

# Call in databook package and set up the environment
library(databook)
data_book <- DataBook$new()

# Filter our data so we are only considering a small portion of it
daily_data <- rpicsa::daily_niger \%>\%
  dplyr::filter(year <= 1950) \%>\%
  dplyr::filter(year > 1945) \%>\%
  dplyr::mutate(year = as.numeric(year)) \%>\%
  dplyr::filter(station_name == "Agades")
data_book$import_data(list(daily_data = daily_data))

 # Run End-Rains:
end_rains(data = "daily_data", date_time = "date", station = "station_name",
          year = "year", rain = "rain",
          start_day = 121, end_day = 300,
          output = "doy", data_book = data_book)
          
 # View output
daily_data_by_station_name_year <- data_book$get_data_frame("daily_data_by_station_name_year")
head(daily_data_by_station_name_year)
}
