% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_binary_file.R
\name{get_binary_file}
\alias{get_binary_file}
\title{Get Binary File}
\usage{
get_binary_file(file_path)
}
\arguments{
\item{file_path}{The path to the file.}
}
\value{
A raw vector containing the binary data of the file.
}
\description{
Reads and returns the binary data of a file.
}
\examples{
# to complete
#pdf_data <- get_binary_file("data/idems_logo.jpg")
#pdf_data <- get_binary_file("data/idems_logo.pdf")

}
