% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_null_data.R
\name{join_null_data}
\alias{join_null_data}
\title{Join Null Data}
\usage{
join_null_data(summary_data, calculated_data)
}
\arguments{
\item{summary_data}{A data frame representing summary data.}

\item{calculated_data}{A data frame containing calculated data.}
}
\value{
A data frame resulting from the full join of \code{summary_data} and \code{calculated_data},
or \code{calculated_data} if \code{summary_data} is NULL.
}
\description{
This function joins two data frames, \code{summary_data} and \code{calculated_data},
using a full join if \code{summary_data} is not NULL. If \code{summary_data} is NULL,
it assigns \code{calculated_data} to \code{summary_data}.
}
\examples{
# summary_data is NULL
summary_data <- NULL
calculated_data <- data.frame(x = 1:5, y = letters[1:5])
join_null_data(summary_data, calculated_data)

# summary_data is not NULL
summary_data <- data.frame(x = 1:3, y = letters[1:3])
calculated_data <- data.frame(x = 4:5, y = letters[4:5])
join_null_data(summary_data, calculated_data)
}
