% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pentad.R
\name{pentad}
\alias{pentad}
\title{Get the pentad component of a date-time object}
\usage{
pentad(date)
}
\arguments{
\item{date}{A date-time object}
}
\value{
a numerical vector of pentad objects corresponding to date variable.
}
\description{
Convert a date or date-time object to a yearly pentad (5-day period)
}
\examples{
pentad(as.Date("2020/12/25"))
pentad(as.Date("1999/01/01"))
}
