% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_start_date_p.R
\name{probability_season_start}
\alias{probability_season_start}
\title{Season start date probabilities}
\usage{
probability_season_start(
  data,
  station = NULL,
  start_rains,
  doy_format = c("doy_366", "doy_365"),
  specified_day
)
}
\arguments{
\item{data}{The data.frame to calculate from that contains the \code{start_rains} variable.}

\item{station}{\code{character(1)} The name of the station column in \code{data}, if the data are for multiple station.}

\item{start_rains}{\code{character(1)} The name of the start of rains column in \code{data}. This can be calculated by the \code{start_rains} function.}

\item{doy_format}{\code{character(1)} Whether \code{start_rains} is given as \code{1-366} (\code{"doy_366"}) or \code{1-365} (\code{"doy_365"}), if \code{doy} is given instead of \code{date}. If a \code{date} is given, this is the value to convert to, and hence, the value that \code{specified_day} takes as well.}

\item{specified_day}{\code{character}. A character vector containing specified doys to calculate the probability of rain before. This is assumed to follow the same as start of rains (see \code{doy_format}).}
}
\value{
Returns a summary table giving the probability that the start of rains will occur on or before the day of years specified.
}
\description{
The probabilities of the start of the rains occurring on or before a set of specified days.
}
\examples{
#TODO#
#x <- start_rains(daily_niger, date_time = "date", station = "station_name", rain = "rain")
#probability_season_start(x, station = "station_name", start_rains = "start_rain", specified_day = c(150, 200))
}
