% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_start_date_p.R
\name{season_start_date_p}
\alias{season_start_date_p}
\title{Season start date probabilities}
\usage{
season_start_date_p(
  data,
  station = NULL,
  start_rains,
  format_SOR = c("366", "365"),
  specified_day,
  format_day = c("date", "doy", "doy_365"),
  date_format = "\%Y-\%m-\%d"
)
}
\arguments{
\item{data}{The data.frame to calculate from that contains the \code{start_rains} variable.}

\item{station}{\code{character(1)} The name of the station column in \code{data}, if the data are for multiple station.}

\item{start_rains}{\code{character(1)} The name of the start of rains column in \code{data}. This can be calculated by the \code{start_rains} function.}

\item{format_SOR}{\code{character(1)} Whether \code{start_rains} is given as \code{1-366} or \code{1-365}, if \code{doy} is given instead of \code{date}.}

\item{specified_day}{\code{\link[base]{Date(1)}} or \code{character(1)}. The specified date or day-of-year to calculate the probability of rain before. If a doy value is given, it is assumed to follow the same format as the value given in \code{format_SOR}.}

\item{format_day}{\code{character(1)} Whether the value given in \code{specified_day} is the \code{date}, \code{doy} (day of year 1-366), or \code{doy_365} (day of year 1-365) value.}

\item{date_format}{\code{character(1)} The date format as used by \code{base::strptime}.}
}
\value{
Returns a summary table giving the probability that the start of rains will occur on or before the date specified.
}
\description{
The probabilities of the start of the rains occurring on or before a set of specified dates.
}
\examples{
#TODO#
#x <- start_rains(daily_niger, date_time = "date", station = "station_name", rain = "rain")
#season_start_date_p(x, station = "station_name", start_rains = "start_rain", specified_day = 150,
#                    format_day = "doy")
}
