% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_length.R
\name{seasonal_length}
\alias{seasonal_length}
\title{Length of the season}
\usage{
seasonal_length(
  summary_data = NULL,
  start_date = NULL,
  end_date = NULL,
  data,
  date_time,
  rain,
  year = NULL,
  station = NULL,
  doy = NULL,
  threshold = 0.85,
  sor_start_day = 1,
  sor_end_day = 366,
  sor_total_rainfall = TRUE,
  sor_over_days = 1,
  sor_amount_rain = 20,
  sor_proportion = FALSE,
  sor_prob_rain_day = 0.8,
  sor_number_rain_days = FALSE,
  sor_min_rain_days = 1,
  sor_rain_day_interval = 2,
  sor_dry_spell = FALSE,
  sor_spell_interval = 21,
  sor_spell_max_dry_days = 9,
  sor_dry_period = FALSE,
  sor_period_interval = 45,
  sor_max_rain = 40,
  sor_period_max_dry_days = 30,
  eos_start_day = 1,
  eos_end_day = 366,
  eos_interval_length = 1,
  eos_min_rainfall = 10
)
}
\arguments{
\item{summary_data}{Summary data frame containing the \code{start_date} and \code{end_date} variables. These variables are calculated from start of rains and end of season functions.
If \code{NULL}, \code{start_date} and \code{end_date} are calculated from the \code{start_of_rains} and \code{end_of_season} functions respectively.}

\item{start_date}{\code{character(1)} The name of the start of rains column in \code{summary_data}. If \code{NULL} it will be created using the \code{start_of_rains} function.}

\item{end_date}{\code{character(1)} The name of the end of season column in \code{summary_data}. If \code{NULL} it will be created using the \code{end_of_seasons} function.}

\item{data}{The data.frame to calculate from.}

\item{date_time}{\code{\link[base]{Date}} The name of the date column in \code{data}.}

\item{rain}{\code{character(1)} The name of the rainfall column in \code{data} to apply the function to.}

\item{year}{\code{character(1)} The name of the year column in \code{data}. If \code{NULL} it will be created using \code{lubridate::year(data[[date_time]])}.}

\item{station}{\code{character(1)} The name of the station column in \code{data}, if the data are for multiple station.}

\item{doy}{\code{character(1)} The name of the day of year column in \code{data} to apply the function to. If \code{NULL} it will be created using the \code{date_time} variable.}

\item{threshold}{\code{numerical(1)} Threshold value for amount (mm) of rainfall in order to count it as a rainy day.}
}
\value{
A data.frame with rainfall summaries for each year in the specified season (between start of the rains and end of season).
}
\description{
Number of days between start of the rains and end of the season
}
\examples{
# Example of season
seasonal_length(data = daily_niger, station = "station_name", date_time = "date",
year = "year", doy = "doy", rain = "rain")

# Using data
start_output <- start_rains(data = daily_niger, station = "station_name", date_time = "date",
                            year = "year", doy = "doy", rain = "rain")
end_output <- end_rains(data = daily_niger, station = "station_name", date_time = "date",
                        year = "year", doy = "doy", rain = "rain")
summary_data <- dplyr::full_join(start_output, end_output)
start_output <- seasonal_length(summary_data = summary_data, start_date = "start_rain", end_date = "end_rain")
}
