% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_dates.R
\name{shift_dates}
\alias{shift_dates}
\title{Shift Dates Function}
\usage{
shift_dates(
  data,
  date,
  s_start_doy,
  doy_shift_name = "s_doy",
  year_shift_name = "s_year"
)
}
\arguments{
\item{data}{A data frame containing date information.}

\item{date}{The name of the date column in the data frame.}

\item{s_start_doy}{The starting day of the year (DOY) for the shift.}

\item{doy_shift_name}{(Optional) The name for the new DOY shift column in the data frame (default is "s_doy").}

\item{year_shift_name}{(Optional) The name for the new year shift column in the data frame (default is "s_year").}
}
\value{
The input data frame with additional columns for DOY shift and year shift.
}
\description{
This function shifts dates in a data frame to a specified day of the year (DOY) and calculates the corresponding year shift.
}
\examples{
data <- data.frame(date = as.Date(c("2023-03-15", "2024-02-29")), year = c(2023, 2024))
shifted_data <- shift_dates(data, "date", 50, "shifted_doy", "shifted_year")

}
\seealso{
\code{\link{yday_366}}: A helper function to calculate day of the year for leap years.
}
