% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spells.R
\name{spells}
\alias{spells}
\title{Running spell length}
\usage{
spells(x, initial_value = NA_real_)
}
\arguments{
\item{x}{A numeric vector}

\item{initial_value}{The initial value of the spell length}
}
\value{
a vector of length \code{x} of the running number of consecutive non-zero values of \code{x}
}
\description{
Calculates the running number of consecutive non-zero values of a vector
}
