% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_rains.R
\name{start_rains}
\alias{start_rains}
\title{Start of the rains}
\usage{
start_rains(
  data,
  date_time,
  station = NULL,
  year = NULL,
  rain = NULL,
  threshold = 0.85,
  doy = NULL,
  start_day = 1,
  end_day = 366,
  total_rainfall = TRUE,
  over_days = 1,
  amount_rain = 20,
  proportion = FALSE,
  prob_rain_day = 0.8,
  number_rain_days = FALSE,
  min_rain_days = 1,
  out_of_rain_days = 2,
  dry_spell = FALSE,
  spell_interval = 21,
  spell_max_dry_days = 9,
  dry_period = FALSE,
  period_interval = 45,
  max_rain = 40,
  period_max_dry_days = 30
)
}
\arguments{
\item{data}{The data.frame to calculate from.}

\item{date_time}{\code{\link[base]{Date}} The name of the date column in \code{data}.}

\item{station}{\code{character(1)} The name of the station column in \code{data}, if the data are for multiple station.}

\item{year}{\code{character(1)} The name of the year column in \code{data}. If \code{NULL} it will be created using \code{lubridate::year(data[[date_time]])}.}

\item{rain}{\code{character(1)} The name of the rainfall column in \code{data} to apply the function to.}

\item{threshold}{\code{numerical(1)} Threshold value for amount (mm) of rainfall in order to count it as a rainy day.}

\item{doy}{\code{character(1)} The name of the day of year column in \code{data} to apply the function to. If \code{NULL} it will be created using the \code{date_time} variable.}

\item{start_day}{\code{numerical(1)} The first day to calculate from in the year (1-366).}

\item{end_day}{\code{numerical(1)} The last day to calculate to in the year (1-366).}

\item{total_rainfall}{}

\item{over_days}{}

\item{amount_rain}{}

\item{proportion}{}

\item{prob_rain_day}{}

\item{number_rain_days}{}

\item{min_rain_days}{}

\item{out_of_rain_days}{}

\item{dry_spell}{}

\item{spell_interval}{}

\item{spell_max_dry_days}{}

\item{dry_period}{}

\item{period_interval}{}

\item{max_rain}{}

\item{period_max_dry_days}{}
}
\value{
A data.frame with the date of the start of the rains for each year
}
\description{
Start of the rains
}
\examples{
#TODO#

# TODO: add in skeleton function explanations
# check against R-Instat function
# read in "spells" function (cdms.products?)
# TODO: option for doy, date, etc
# TODO: checkmate:: functions
}
