% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_calculation.R
\name{summary_calculation}
\alias{summary_calculation}
\title{Summary Calculation}
\usage{
summary_calculation(
  data,
  date_time,
  year = NULL,
  month = NULL,
  station = NULL,
  columns_to_summarise,
  to = c("annual", "monthly"),
  summaries = c("mean", "min", "max", "sum"),
  na_rm = FALSE,
  na_prop = NULL,
  na_n = NULL,
  na_consec = NULL,
  na_n_non = NULL,
  data_book = NULL
)
}
\arguments{
\item{data}{The data.frame to calculate from.}

\item{date_time}{\code{\link[base]{Date}} The name of the date column in \code{data}.}

\item{year}{\code{character(1)} The name of the year column in \code{data}. If \code{NULL} it will be created using \code{lubridate::year(data[[date_time]])}.}

\item{month}{\code{character(1)} The name of the month column in \code{data}. If \code{NULL} it will be created using \code{lubridate::month(data[[date_time]])}.}

\item{station}{\code{character(1)} The name of the station column in \code{data}, if the data are for multiple station.}

\item{columns_to_summarise}{\code{character(1)} The name of the columns to summarise in \code{data}.}

\item{to}{\code{character(1)} Default \code{annual}. The period of time to calculate the mean temperature columns over (options are \code{annual} or \code{monthly}).}

\item{summaries}{\code{character} The summaries to display. Options are \code{"mean"}, \code{"max"}, \code{"min"}, \code{"sum"}}

\item{na_rm}{\code{logical(1)}. Should missing values (including \code{NaN}) be removed?}

\item{na_prop}{\code{integer(1)} Max proportion of missing values allowed}

\item{na_n}{\code{integer(1)} Max number of missing values allowed}

\item{na_consec}{\code{integer(1)} Max number of consecutive missing values allowed}

\item{na_n_non}{\code{integer(1)} Min number of non-missing values required}

\item{data_book}{The data book object where the data object is stored, default \code{NULL}.}
}
\value{
A data.frame with summaries for the columns_to_summarise specified by year and/or month (and optionally station)
}
\description{
Returns a summary data frame giving either the mean, minimum, and/or maximum temperatures by year and/or month, and by optionally station.
}
