% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual_rain.R
\name{annual_rain}
\alias{annual_rain}
\title{Annual total rainfall}
\usage{
annual_rain(
  data,
  year = NULL,
  station = NULL,
  rain,
  total_rain = TRUE,
  n_rain = TRUE,
  rain_day = 0.85,
  na_rm = FALSE,
  na_prop = NULL,
  na_n = NULL,
  na_consec = NULL,
  na_n_non = NULL
)
}
\arguments{
\item{data}{The data.frame to calculate from.}

\item{year}{\code{character(1)} The name of the year column in \code{data}. If \code{NULL} it will be created using \code{lubridate::year(data[[date_time]])}.}

\item{station}{\code{character(1)} The name of the station column in \code{data}, if the data are for multiple station.}

\item{rain}{\code{character(1)} The name of the rainfall column in \code{data} to apply the function to.}

\item{total_rain}{\code{logical(1)} default \code{TRUE}. Display the total rainfall value for each year.}

\item{n_rain}{\code{logical(1)} default \code{TRUE}. Display the number of rainfall days.}

\item{rain_day}{\code{numerical(1)} If \code{n_rain = TRUE}, the minimum rainfall value in a day for that day to count as a rainfall day.}

\item{na_rm}{\code{logical(1)}. Should missing values (including \code{NaN}) be removed?}

\item{na_prop}{\code{integer(1)} Max proportion of missing values allowed}

\item{na_n}{\code{integer(1)} Max number of missing values allowed}

\item{na_consec}{\code{integer(1)} Max number of consecutive missing values allowed}

\item{na_n_non}{\code{integer(1)} Min number of non-missing values required}
}
\value{
A data.frame with rainfall summaries for each year
}
\description{
Returns a summary data frame giving the total rainfall each year from 1 Jan to 31 Dec.
}
\examples{
#daily_niger_1 <- daily_niger \%>\% dplyr::filter(year > 1960)
#annual_rain(data = daily_niger, date_time  = "date", station = "station_name",
#            rain = "rain", na_prop = 0.9)
}
