% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crops_definitions.R
\name{crops_definitions}
\alias{crops_definitions}
\title{Crop Definitions}
\usage{
crops_definitions(
  data,
  date_time,
  station = NULL,
  rain,
  year = NULL,
  doy = NULL,
  water_requirements,
  planting_dates,
  crop_length,
  start_check = TRUE,
  season_data = NULL,
  start_day,
  end_day
)
}
\arguments{
\item{data}{The data.frame containing the rainfall data.}

\item{date_time}{\code{\link[base]{Date}} The name of the date column in \code{data}.}

\item{station}{\code{character(1)} The name of the station column in \code{data}, if the data are for multiple station.}

\item{rain}{\code{character(1)} The name of the rainfall column in \code{data} to apply the function to.}

\item{year}{\code{character(1)} The name of the year column in \code{data}. If \code{NULL} it will be created using \code{lubridate::year(data[[date_time]])}.}

\item{doy}{\code{character(1)} The name of the day of year column in \code{data}. If \code{NULL} it will be created using \code{lubridate::year(data[[date_time]])}.}

\item{water_requirements}{\code{numeric} Vector containing water requirements.}

\item{planting_dates}{\code{numeric} Vector containing planting dates requirements.}

\item{crop_length}{\code{numeric} Vector containing seasonal crop length requirements.}

\item{start_check}{\code{logical} A logical value indicating whether to check the start day condition (default is \code{TRUE}).}

\item{season_data}{The data frame containing the seasonal data.}

\item{start_day}{\code{character(1)} The name of the column in the season_data data frame that represents the start day. This can be calculated prior to using this function by the \code{start_rains} function.}

\item{end_day}{\code{character(1)} The name of the column in the season_data data frame that represents the end day. This should be calculated prior to using this function by the \code{end_rains} function.}
}
\value{
TODO
}
\description{
Calculate the probabilities of crop success for given planting maturity lengths, seasonal total rainfall requirements, and planting dates. It is not required that the start of the rains occurs on or before the planting date.
}
\examples{
#TODO
}
