% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end_season.R
\name{end_season}
\alias{end_season}
\title{End of the rains}
\usage{
end_season(
  data,
  date_time,
  station = NULL,
  year = NULL,
  rain = NULL,
  doy = NULL,
  start_day = 1,
  end_day = 366,
  output = c("doy", "date", "both"),
  capacity = 60,
  water_balance_max = 0.5,
  evaporation = c("value", "variable"),
  evaporation_value = 5,
  evaporation_variable = NULL
)
}
\arguments{
\item{data}{The data.frame to calculate from.}

\item{date_time}{\code{\link[base]{Date}} The name of the date column in \code{data}.}

\item{station}{\code{character(1)} The name of the station column in \code{data}, if the data are for multiple station.}

\item{year}{\code{character(1)} The name of the year column in \code{data}. If \code{NULL} it will be created using \code{lubridate::year(data[[date_time]])}.}

\item{rain}{\code{character(1)} The name of the rainfall column in \code{data} to apply the function to.}

\item{doy}{\code{character(1)} The name of the day of year column in \code{data} to apply the function to. If \code{NULL} it will be created using the \code{date_time} variable.}

\item{start_day}{\code{numerical(1)} The first day to calculate from in the year (1-366).}

\item{end_day}{\code{numerical(1)} The last day to calculate to in the year (1-366).}

\item{output}{\code{character(1)} Whether to give the start of rains by day of year (doy), date, or both. Default \code{"doy"}.}

\item{capacity}{\code{numerical(1)} Water capacity of the soil (default \code{60}).}

\item{water_balance_max}{\code{numerical(1)} Maximum water balance value (default \code{0.5}).}

\item{evaporation}{\code{character(1)} Whether to give evaporation as a value or variable. Default \code{"value"}.}

\item{evaporation_value}{\code{numerical(1)} If \code{evaporation = "value"}, the numerical value of amount of evaporation per day (default \code{5}).}

\item{evaporation_variable}{\code{character(1)} If \code{evaporation = "variable"}, the variable in \code{data} that corresponds to the evaporation column.}
}
\value{
A data.frame with the day of year and/or date for the end of the season for each year (and station).
}
\description{
Last occurrence from day q with at least z mm rainfall on a set of days.
}
\examples{
#TODO # is same as R-Instat
# end_season(niger_daily_1, date_time = "date", year = "year",
#rain = "rain", doy = "doy", output = "doy",
#capacity = 60, water_balance_max = 0.5,
#evaporation_value = 5)
}
