% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extremes.R
\name{get_extremes}
\alias{get_extremes}
\title{Get Extreme Data}
\usage{
get_extremes(
  data,
  station = NULL,
  year,
  element,
  type = c("percentile", "threshold"),
  value = 95,
  direction = c("greater", "less")
)
}
\arguments{
\item{data}{A data frame containing the data to be analysed.}

\item{station}{The name of the \code{station} column in 'data'. Default \code{NULL}.}

\item{year}{The name of the \code{year} column in 'data'.}

\item{element}{The name of the column in 'data' for which extremes are to be found.}

\item{type}{A character string specifying the mode of operation. It can be either \code{"percentile"} or \code{"threshold"}. Here, \code{"percentile"} identifies values above a certain percentile (e.g., 95th percentile); \code{"threshold"} identifies values above a specific threshold value.}

\item{value}{A numeric value specifying the percentile or threshold, depending on the 'type' parameter. If \code{type == "percentile"}, \code{value} is the percentile (e.g., 95 for 95th percentile). If \code{type == "threshold"}, \code{value} is the threshold value (e.g., 50 mm for rainfall).}

\item{direction}{A character string specifying the direction for the operation. It can be either \code{"greater"} or \code{"less"}.}
}
\value{
A filtered data frame where the \code{element} values are considered extreme based on the specified \code{type} and \code{value}.
}
\description{
This function identifies extreme values in a specified element (column) of a data frame. It can operate in two modes: percentile-based and threshold-based.
}
\examples{
# data(daily_niger)
# filtered_data <- get_extremes(data = daily_niger, station = NULL, year = "year", element = "rain", type = "threshold", value = 50)
}
