% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_temp.R
\name{mean_temperature}
\alias{mean_temperature}
\title{Mean Temperature (Month or annually)}
\usage{
mean_temperature(
  data,
  date_time,
  tmin = NULL,
  tmax = NULL,
  year = NULL,
  month = NULL,
  station = NULL,
  to = c("annual", "monthly"),
  na_rm = FALSE,
  na_prop = NULL,
  na_n = NULL,
  na_consec = NULL,
  na_n_non = NULL
)
}
\arguments{
\item{data}{The data.frame to calculate from.}

\item{date_time}{\code{\link[base]{Date}} The name of the date column in \code{data}.}

\item{tmin}{\code{character(1)} The name of the minimum temperature column in \code{data} to apply the function to.}

\item{tmax}{\code{character(1)} The name of the maximum temperature column in \code{data} to apply the function to.}

\item{year}{\code{character(1)} The name of the year column in \code{data}. If \code{NULL} it will be created using \code{lubridate::year(data[[date_time]])}.}

\item{month}{\code{character(1)} The name of the month column in \code{data}. If \code{NULL} it will be created using \code{lubridate::month(data[[date_time]])}.}

\item{station}{\code{character(1)} The name of the station column in \code{data}, if the data are for multiple station.}

\item{to}{\code{character(1)} Default \code{annual}. The period of time to calculate the mean temperature columns over (options are \code{annual} or \code{monthly}).}

\item{na_rm}{\code{logical(1)}. Should missing values (including \code{NaN}) be removed?}

\item{na_prop}{\code{integer(1)} Max proportion of missing values allowed}

\item{na_n}{\code{integer(1)} Max number of missing values allowed}

\item{na_consec}{\code{integer(1)} Max number of consecutive missing values allowed}

\item{na_n_non}{\code{integer(1)} Min number of non-missing values required}
}
\value{
A data.frame with mean summaries for each year or year and month for the minimum daily temperature and/or the maximum daily temperature.
}
\description{
: Returns a summary data frame giving either the mean minimum and/or maximum temperature each year from 1 Jan to 31 Dec, or by year and month.
}
\examples{
#daily_niger_1 <- daily_niger \%>\% filter(year < 1950)
#mean_temperature(data = daily_niger_1, date_time  = "date", station = "station_name",
#             tmax = "tmax", tmin = "tmin", na_prop = 0.05)
}
