% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_length.R
\name{seasonal_length}
\alias{seasonal_length}
\title{Length of the season}
\usage{
seasonal_length(
  summary_data,
  start_date,
  end_date,
  start_rain_status = NULL,
  end_rain_status = NULL,
  season_length_save_name = "length",
  occurrence_save_name = "occurrence",
  data_book = NULL
)
}
\arguments{
\item{summary_data}{Summary data frame containing the \code{start_date} and \code{end_date} variables. These variables are calculated from start of rains and end of season functions.}

\item{start_date}{\code{character(1)} The name of the start of rains column in \code{summary_data}. If \code{NULL} it will be created using the \code{start_of_rains} function.}

\item{end_date}{\code{character(1)} The name of the end of season column in \code{summary_data}. If \code{NULL} it will be created using the \code{end_of_seasons} function.}

\item{start_rain_status}{\code{character(1)} The name of the variable that indicates the start of rain status}

\item{end_rain_status}{\code{character(1)} The name of the variable that indicates the end of rain status}

\item{season_length_save_name}{\code{character(1)} The name used to save the length of season}

\item{occurrence_save_name}{\code{character(1)} The name used to save the status results}

\item{data_book}{The data book object where the data object is stored, default \code{NULL}.}
}
\value{
A data.frame with length of rainfall season for each year in the specified season (between start of the rains and end of season).
}
\description{
Number of days between start of the rains and end of the season
}
\examples{
# Example: First calculate start and end rains
data_book <- DataBook$new()
daily_data <- rpicsa::daily_niger \%>\%
  dplyr::filter(year <= 1950) \%>\%
  dplyr::filter(year > 1945) \%>\%
  dplyr::mutate(year = as.numeric(year)) \%>\%
  dplyr::filter(station_name == "Agades")
data_book$import_data(list(daily_data = daily_data))

start_rains(data = "daily_data",
            date_time = "date",
            station = "station_name",
            year = "year",
            rain = "rain",
            start_day = 121,
            end_day = 300,
            output = c("doy", "status"),
            data_book = data_book)

end_rains(data = "daily_data",
          date_time = "date",
          station = "station_name",
          year = "year",
          rain = "rain",
          start_day = 121,
          end_day = 300,
          output = c("doy", "status"),
          data_book = data_book)

# then run the summaries for seasonal rainfall
seasonal_length(summary_data = "daily_data_by_station_name_year",
                start_date = "start_rain",
                end_date = "end_rains",
                start_rain_status = "start_rain_status",
                end_rain_status = "end_rains_status",
                data_book = data_book
)
daily_data_by_station_name_year <- data_book$get_data_frame("daily_data_by_station_name_year")
daily_data_by_station_name_year 
}
