% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_length.R
\name{seasonal_length}
\alias{seasonal_length}
\title{Length of the season}
\usage{
seasonal_length(
  summary_data = NULL,
  start_date = NULL,
  end_date = NULL,
  s_start_month = 1,
  data = NULL,
  date_time = NULL,
  rain = NULL,
  year = NULL,
  station = NULL,
  doy = NULL,
  threshold = 0.85,
  sor_start_day = 1,
  sor_end_day = 366,
  sor_total_rainfall_comparison = c("amount", "proportion", "evaporation"),
  sor_over_days = 1,
  sor_amount_rain = 20,
  sor_prob_rain_day = 0.8,
  sor_evaporation_variable = NULL,
  sor_fraction = 0.5,
  sor_number_rain_days = FALSE,
  sor_min_rain_days = 1,
  sor_rain_day_interval = 2,
  sor_dry_spell = FALSE,
  sor_spell_interval = 21,
  sor_spell_max_dry_days = 9,
  sor_dry_period = FALSE,
  sor_period_interval = 45,
  sor_max_rain = 40,
  sor_period_max_dry_days = 30,
  end_type = c("season", "rains"),
  eos_start_day = 1,
  eos_end_day = 366,
  eor_interval_length = 1,
  eor_min_rainfall = 10,
  eos_capacity = 60,
  eos_water_balance_max = 0.5,
  eos_evaporation = c("value", "variable"),
  eos_evaporation_value = 5,
  eos_evaporation_variable = NULL,
  data_book,
  season_length_save_name = "length",
  occurrence_save_name = "length_status",
  start_rain_status = NULL,
  end_rain_status = NULL
)
}
\arguments{
\item{summary_data}{Summary data frame containing the \code{start_date} and \code{end_date} variables. These variables are calculated from start of rains and end of season functions.
If \code{NULL}, \code{start_date} and \code{end_date} are calculated from the \code{start_of_rains} and \code{end_of_season} functions respectively.}

\item{start_date}{\code{character(1)} The name of the start of rains column in \code{summary_data}. If \code{NULL} it will be created using the \code{start_of_rains} function.}

\item{end_date}{\code{character(1)} The name of the end of season column in \code{summary_data}. If \code{NULL} it will be created using the \code{end_of_seasons} function.}

\item{s_start_month}{\code{integer(1)} Month (1–12) to treat as the start of the “year” when creating \code{year} or \code{doy}. Default \code{NULL} (assumes January).}

\item{data}{The daily data.frame to calculate rainfall from.}

\item{date_time}{\code{\link[base]{Date}} The name of the date column in \code{data}.}

\item{rain}{\code{character(1)} The name of the rainfall column in \code{data} to apply the function to.}

\item{year}{\code{character(1)} The name of the year column in \code{data}. If \code{NULL} it will be created using \code{lubridate::year(data[[date_time]])}.}

\item{station}{\code{character(1)} The name of the station column in \code{data}, if the data are for multiple station.}

\item{doy}{\code{character(1)} The name of the day of year column in \code{data} to apply the function to. If \code{NULL} it will be created using the \code{date_time} variable.}

\item{threshold}{\code{numerical(1)} threshold value for amount (mm) of rainfall in order to count it as a rainy day.}

\item{sor_start_day}{\code{numerical(1)} The first day to calculate from in the year (1-366).}

\item{sor_end_day}{\code{numerical(1)} The last day to calculate to in the year (1-366).}

\item{sor_total_rainfall_comparison}{Method for multi‐day definition: \code{"amount"}, \code{"proportion"}, or \code{"evaporation"}. Default \code{"amount"}.}

\item{sor_over_days}{\code{numerical(1)} Only works if \code{total_rainfall = TRUE}. This is the number of days to total the rainfall over.}

\item{sor_amount_rain}{\code{numerical(1)} If \code{sor_total_rainfall_comparison = "amount"}, the amount of rainfall to expect over the period defined in \code{over_days}.}

\item{sor_prob_rain_day}{\code{numerical(1)} Only works if \code{sor_total_rainfall_comparison = "proportion"} This is the proportion.}

\item{sor_evaporation_variable}{\code{character(1)} Only works if \code{sor_total_rainfall_comparison = "evaporation"}. Name of evaporation (mm) column, if using \code{"evaporation"}. Default \code{NULL}.}

\item{sor_fraction}{\code{numeric(1)} Multiplier (0–1) applied to evaporation values when \code{total_rainfall_comparison="evaporation"}. Default \code{0.5}.}

\item{sor_number_rain_days}{\code{logical(1)} default \code{FALSE}. If \code{TRUE}, define start of the rains by the number of rainy days (\code{min_rain_days}) over a period. The period is given in days in the \code{rain_day_interval} parameter.}

\item{sor_min_rain_days}{\code{numerical(1)} Only if \code{number_rain_days = TRUE}. This is the minimum number of rainy days to define start of rains in a given period. The period is given in days in the \code{rain_day_interval} parameter.}

\item{sor_rain_day_interval}{\code{numerical(1)} Only if \code{number_rain_days = TRUE}, the interval in days that the \code{number_rain_days} is defined in.}

\item{sor_dry_spell}{\code{logical(1)} default \code{FALSE}. If \code{TRUE}, define start of the rains by a maximum number of dry days (\code{spell_max_dry_days}) over a given period of days (\code{spell_interval}).}

\item{sor_spell_interval}{\code{numerical(1)} Only if \code{dry_spell = TRUE}, the interval in days that the \code{dry_spell} is defined in.}

\item{sor_spell_max_dry_days}{\code{numerical(1)} Only if \code{dry_spell = TRUE}. This is the maximum number of dry days to define start of rains in a given period. The period is given in days in the \code{spell_interval} parameter.}

\item{sor_dry_period}{\code{logical(1)} default \code{FALSE}. If \code{TRUE}, define start of the rains by the maximum rain and maximum dry days in a given interval. The maximum rainfall amount is given in the \code{max_rain} parameter, the maximum dry days is given in the \code{period_max_dry_days} parameter, and the interval length is given in the \code{period_interval} parameter.}

\item{sor_period_interval}{\code{numerical(1)} Only if \code{dry_period = TRUE}, the interval in days that the \code{dry_period} is defined in.}

\item{sor_max_rain}{\code{numerical(1)} Only if \code{dry_period = TRUE}, the maximum rainfall to occur in a given period.}

\item{sor_period_max_dry_days}{\code{numerical(1)} Only if \code{dry_period = TRUE}. the maximum period of dry days to occur in a given period.}

\item{end_type}{\code{character(1)} If \code{is.null(end_date)}, \code{end_type} is whether the end of seasons or end of rains is used. Options are c(\verb{"season", "rains"}), default \code{"season"}.}

\item{eos_start_day}{\code{numerical(1)} The first day to calculate from in the year (1-366).}

\item{eos_end_day}{\code{numerical(1)} The last day to calculate to in the year (1-366).}

\item{eor_interval_length}{\code{numerical(1)} Number of days for the minimum rainfall to fall in.}

\item{eor_min_rainfall}{\code{numerical(1)} Minimum amount of rainfall to occur on the set of days  defined in \code{interval_length}.}

\item{eos_capacity}{\code{numerical(1)} Water capacity of the soil (default \code{60}).}

\item{eos_water_balance_max}{\code{numerical(1)} Maximum water balance value (default \code{0.5}).}

\item{eos_evaporation}{\code{character(1)} Whether to give evaporation as a value or variable. Default \code{"value"}.}

\item{eos_evaporation_value}{\code{numerical(1)} If \code{evaporation = "value"}, the numerical value of amount of evaporation per day (default \code{5}).}

\item{eos_evaporation_variable}{\code{character(1)} If \code{evaporation = "variable"}, the variable in \code{data} that corresponds to the evaporation column.}

\item{data_book}{The data book object where the data object is stored, default \code{NULL}.}

\item{season_length_save_name}{\code{character(1)} The name used to save the length of season}

\item{occurrence_save_name}{\code{character(1)} The name used to save the status results}

\item{start_rain_status}{\code{character(1)} The name of the variable that indicates the end of rain status}
}
\value{
A data.frame with length of rainfall season for each year in the specified season (between start of the rains and end of season).
}
\description{
Number of days between start of the rains and end of the season
}
\examples{
# TODO

}
