% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_rain.R
\name{seasonal_rain}
\alias{seasonal_rain}
\title{Seasonal total rainfall}
\usage{
seasonal_rain(
  data,
  date_time,
  rain,
  year = NULL,
  station = NULL,
  summary_data = NULL,
  start_date = NULL,
  end_date = NULL,
  total_rain = TRUE,
  n_rain = TRUE,
  rain_day = 0.85,
  na_rm = FALSE,
  na_prop = NULL,
  na_n = NULL,
  na_consec = NULL,
  na_n_non = NULL
)
}
\arguments{
\item{data}{The data.frame to calculate from.}

\item{date_time}{\code{\link[base]{Date}} The name of the date column in \code{data}.}

\item{rain}{\code{character(1)} The name of the rainfall column in \code{data} to apply the function to.}

\item{year}{\code{character(1)} The name of the year column in \code{data}. If \code{NULL} it will be created using \code{lubridate::year(data[[date_time]])}.}

\item{station}{\code{character(1)} The name of the station column in \code{data}, if the data are for multiple station.}

\item{summary_data}{Summary data frame containing the \code{start_date} and \code{end_date} variables. These variables are calculated from start of rains and end of season functions.
If \code{NULL}, \code{start_date} and \code{end_date} are calculated from the \code{start_of_rains} and \code{end_of_season} functions respectively.}

\item{start_date}{\code{character(1)} The name of the start of rains column in \code{summary_data}. If \code{NULL} it will be created using the \code{start_of_rains} function.}

\item{end_date}{\code{character(1)} The name of the end of season column in \code{summary_data}. If \code{NULL} it will be created using the \code{end_of_seasons} function.}

\item{total_rain}{\code{logical(1)} default \code{TRUE}. Display the total rainfall value for each year.}

\item{n_rain}{\code{logical(1)} default \code{TRUE}. Display the number of rainfall days.}

\item{rain_day}{\code{numerical(1)} If \code{n_rain = TRUE}, the minimum rainfall value in a day for that day to count as a rainfall day.}

\item{na_rm}{\code{logical(1)}. Should missing values (including \code{NaN}) be removed?}

\item{na_prop}{\code{integer(1)} Max proportion of missing values allowed}

\item{na_n}{\code{integer(1)} Max number of missing values allowed}

\item{na_consec}{\code{integer(1)} Max number of consecutive missing values allowed}

\item{na_n_non}{\code{integer(1)} Min number of non-missing values required}
}
\value{
A data.frame with rainfall summaries for each year in the specified season (between start of the rains and end of season).
}
\description{
Total annual rainfall between start of the rains and end of the season.
}
\examples{
# TODO
}
