% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_rain.R
\name{seasonal_rain}
\alias{seasonal_rain}
\title{Seasonal total rainfall}
\usage{
seasonal_rain(
  summary_data = NULL,
  start_date = NULL,
  end_date = NULL,
  data,
  date_time,
  year = NULL,
  station = NULL,
  doy = NULL,
  rain = NULL,
  total_rain = TRUE,
  n_rain = TRUE,
  rain_day = 0.85,
  na_rm = FALSE,
  na_prop = NULL,
  na_n = NULL,
  na_consec = NULL,
  na_n_non = NULL,
  threshold = 0.85,
  sor_start_day = 1,
  sor_end_day = 366,
  sor_total_rainfall = TRUE,
  sor_over_days = 1,
  sor_amount_rain = 20,
  sor_proportion = FALSE,
  sor_prob_rain_day = 0.8,
  sor_number_rain_days = FALSE,
  sor_min_rain_days = 1,
  sor_rain_day_interval = 2,
  sor_dry_spell = FALSE,
  sor_spell_interval = 21,
  sor_spell_max_dry_days = 9,
  sor_dry_period = FALSE,
  sor_period_interval = 45,
  sor_max_rain = 40,
  sor_period_max_dry_days = 30,
  eos_start_day = 1,
  eos_end_day = 366,
  eos_interval_length = 1,
  eos_min_rainfall = 10
)
}
\arguments{
\item{summary_data}{Summary data frame containing the \code{start_date} and \code{end_date} variables. These variables are calculated from start of rains and end of season functions.
If \code{NULL}, \code{start_date} and \code{end_date} are calculated from the \code{start_of_rains} and \code{end_of_season} functions respectively.}

\item{start_date}{\code{character(1)} The name of the start of rains column in \code{summary_data}. If \code{NULL} it will be created using the \code{start_of_rains} function.}

\item{end_date}{\code{character(1)} The name of the end of season column in \code{summary_data}. If \code{NULL} it will be created using the \code{end_of_seasons} function.}

\item{data}{The daily data.frame to calculate rainfall from.}

\item{date_time}{\code{\link[base]{Date}} The name of the date column in \code{data}.}

\item{year}{\code{character(1)} The name of the year column in \code{data}. If \code{NULL} it will be created using \code{lubridate::year(data[[date_time]])}.}

\item{station}{\code{character(1)} The name of the station column in \code{data}, if the data are for multiple station.}

\item{doy}{\code{character(1)} The name of the day of year column in \code{data} to apply the function to. If \code{NULL} it will be created using the \code{date_time} variable.}

\item{rain}{\code{character(1)} The name of the rainfall column in \code{data} to apply the function to.}

\item{total_rain}{\code{logical(1)} default \code{TRUE}. Display the total rainfall value for each year.}

\item{n_rain}{\code{logical(1)} default \code{TRUE}. Display the number of rainfall days.}

\item{rain_day}{\code{numerical(1)} If \code{n_rain = TRUE}, the minimum rainfall value in a day for that day to count as a rainfall day.}

\item{na_rm}{\code{logical(1)}. Should missing values (including \code{NaN}) be removed?}

\item{na_prop}{\code{integer(1)} Max proportion of missing values allowed}

\item{na_n}{\code{integer(1)} Max number of missing values allowed}

\item{na_consec}{\code{integer(1)} Max number of consecutive missing values allowed}

\item{na_n_non}{\code{integer(1)} Min number of non-missing values required}

\item{sor_start_day}{\code{numerical(1)} The first day to calculate from in the year (1-366).}

\item{sor_end_day}{\code{numerical(1)} The last day to calculate to in the year (1-366).}

\item{sor_total_rainfall}{\code{logical(1)} default \code{TRUE}. Start of the rains to be defined by the total or proportion of rainfall over a period.}

\item{sor_over_days}{\code{numerical(1)} Only works if \code{total_rainfall = TRUE}. This is the number of days to total the rainfall over.}

\item{sor_amount_rain}{\code{numerical(1)} If \code{total_rainfall = TRUE} and \code{proportion = FALSE}, the amount of rainfall to expect over the period defined in \code{over_days}.}

\item{sor_proportion}{\code{logical(1)} default \code{FALSE}, only valid if \code{total_rainfall = TRUE}. If \code{TRUE}, Start of the rains to be defined by proportion of rainfall over a period. This proportion is given in \code{prob_rain_day}. Otherwise, defined by the amount of rainfall over a period. The amount is given in \code{amount_rain}.}

\item{sor_prob_rain_day}{\code{numerical(1)} Only works if \code{total_rainfall = TRUE} and \code{proportion = TRUE} This is the number}

\item{sor_number_rain_days}{\code{logical(1)} default \code{FALSE}. If \code{TRUE}, define start of the rains by the number of rainy days (\code{min_rain_days}) over a period. The period is given in days in the \code{rain_day_interval} parameter.}

\item{sor_min_rain_days}{\code{numerical(1)} Only if \code{number_rain_days = TRUE}. This is the minimum number of rainy days to define start of rains in a given period. The period is given in days in the \code{rain_day_interval} parameter.}

\item{sor_rain_day_interval}{\code{numerical(1)} Only if \code{number_rain_days = TRUE}, the interval in days that the \code{number_rain_days} is defined in.}

\item{sor_dry_spell}{\code{logical(1)} default \code{FALSE}. If \code{TRUE}, define start of the rains by a maximum number of dry days (\code{spell_max_dry_days}) over a given period of days (\code{spell_interval}).}

\item{sor_spell_interval}{\code{numerical(1)} Only if \code{dry_spell = TRUE}, the interval in days that the \code{dry_spell} is defined in.}

\item{sor_spell_max_dry_days}{\code{numerical(1)} Only if \code{dry_spell = TRUE}. This is the maximum number of dry days to define start of rains in a given period. The period is given in days in the \code{spell_interval} parameter.}

\item{sor_dry_period}{\code{logical(1)} default \code{FALSE}. If \code{TRUE}, define start of the rains by the maximum rain and maximum dry days in a given interval. The maximum rainfall amount is given in the \code{max_rain} parameter, the maximum dry days is given in the \code{period_max_dry_days} parameter, and the interval length is given in the \code{period_interval} parameter.}

\item{sor_period_interval}{\code{numerical(1)} Only if \code{dry_period = TRUE}, the interval in days that the \code{dry_period} is defined in.}

\item{sor_max_rain}{\code{numerical(1)} Only if \code{dry_period = TRUE}, the maximum rainfall to occur in a given period.}

\item{sor_period_max_dry_days}{\code{numerical(1)} Only if \code{dry_period = TRUE}. the maximum period of dry days to occur in a given period.}

\item{eos_start_day}{\code{numerical(1)} The first day to calculate from in the year (1-366).}

\item{eos_end_day}{\code{numerical(1)} The last day to calculate to in the year (1-366).}

\item{eos_interval_length}{\code{numerical(1)} Number of days for the minimum rainfall to fall in.}

\item{eos_min_rainfall}{\code{numerical(1)} Minimum amount of rainfall to occur on the set of days defined in \code{interval_length}.}

\item{sor_output}{\code{character(1)} Whether to give the start of rains by day of year (doy), date, or both. Default \code{"doy"}.}
}
\value{
A data.frame with rainfall summaries for each year in the specified season (between start of the rains and end of season).
}
\description{
Total annual rainfall between start of the rains and end of the season.
}
\examples{
# TODO
library(cdms.products)
data(daily_niger)

seasonal_rain(data = daily_niger, station = "station_name", date_time = "date",
year = "year", doy = "doy", rain = "rain")

# Using data
start_output <- start_rains(data = daily_niger, station = "station_name", date_time = "date",
                            year = "year", doy = "doy", rain = "rain")
end_output <- end_rains(data = daily_niger, station = "station_name", date_time = "date",
                        year = "year", doy = "doy", rain = "rain")
summary_data <- dplyr::full_join(start_output, end_output)
start_output <- seasonal_rain(summary_data = summary_data, date_time = "date", station = "station_name", data = daily_niger, year = "year", start_date = "start_rain", end_date = "end_rain", rain = "rain")

# Or to create one of the start rains in the function:
summary_data <- end_output
start_output <- seasonal_rain(summary_data = summary_data, date_time = "date", station = "station_name", data = daily_niger, year = "year", end_date = "end_rain", rain = "rain", threshold = 20)
}
