% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_rains.R
\name{start_rains}
\alias{start_rains}
\title{Start of the rains}
\usage{
start_rains()
}
\value{
A data.frame with the date of the start of the rains for each year
}
\description{
Start of the rains
}
\examples{
#TODO
}
