% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_rains.R
\name{start_rains}
\alias{start_rains}
\title{Start of the rains}
\usage{
start_rains(
  data,
  date_time,
  station = NULL,
  year = NULL,
  rain = NULL,
  threshold = 0.85,
  doy = NULL,
  start_day = 1,
  end_day = 366,
  s_start_doy = NULL,
  drop = TRUE,
  output = c("doy", "date", "both"),
  total_rainfall = TRUE,
  over_days = 1,
  amount_rain = 20,
  proportion = FALSE,
  prob_rain_day = 0.8,
  number_rain_days = FALSE,
  min_rain_days = 1,
  rain_day_interval = 2,
  dry_spell = FALSE,
  spell_interval = 21,
  spell_max_dry_days = 9,
  dry_period = FALSE,
  period_interval = 45,
  max_rain = 40,
  period_max_dry_days = 30
)
}
\arguments{
\item{data}{The data.frame to calculate from.}

\item{date_time}{\code{\link[base]{Date}} The name of the date column in \code{data}.}

\item{station}{\code{character(1)} The name of the station column in \code{data}, if the data are for multiple station.}

\item{year}{\code{character(1)} The name of the year column in \code{data}. If \code{NULL} it will be created using \code{lubridate::year(data[[date_time]])}.}

\item{rain}{\code{character(1)} The name of the rainfall column in \code{data} to apply the function to.}

\item{threshold}{\code{numerical(1)} threshold value for amount (mm) of rainfall in order to count it as a rainy day.}

\item{doy}{\code{character(1)} The name of the day of year column in \code{data} to apply the function to. If \code{NULL} it will be created using the \code{date_time} variable.}

\item{start_day}{\code{numerical(1)} The first day to calculate from in the year (1-366).}

\item{end_day}{\code{numerical(1)} The last day to calculate to in the year (1-366).}

\item{s_start_doy}{\code{numerical(1)} Default \code{NULL} (if \code{NULL}, \code{s_start_doy = 1}. The day of year to state is the first day of year.}

\item{drop}{\code{logical(1)} default \code{TRUE}. Whether to drop years where there are \code{NA} data for the rainfall.}

\item{output}{\code{character(1)} Whether to give the start of rains by day of year (doy), date, or both. Default \code{"doy"}.}

\item{total_rainfall}{\code{logical(1)} default \code{TRUE}. Start of the rains to be defined by the total or proportion of rainfall over a period.}

\item{over_days}{\code{numerical(1)} Only works if \code{total_rainfall = TRUE}. This is the number of days to total the rainfall over.}

\item{amount_rain}{\code{numerical(1)} If \code{total_rainfall = TRUE} and \code{proportion = FALSE}, the amount of rainfall to expect over the period defined in \code{over_days}.}

\item{proportion}{\code{logical(1)} default \code{FALSE}, only valid if \code{total_rainfall = TRUE}. If \code{TRUE}, Start of the rains to be defined by proportion of rainfall over a period. This proportion is given in \code{prob_rain_day}. Otherwise, defined by the amount of rainfall over a period. The amount is given in \code{amount_rain}.}

\item{prob_rain_day}{\code{numerical(1)} Only works if \code{total_rainfall = TRUE} and \code{proportion = TRUE} This is the number}

\item{number_rain_days}{\code{logical(1)} default \code{FALSE}. If \code{TRUE}, define start of the rains by the number of rainy days (\code{min_rain_days}) over a period. The period is given in days in the \code{rain_day_interval} parameter.}

\item{min_rain_days}{\code{numerical(1)} Only if \code{number_rain_days = TRUE}. This is the minimum number of rainy days to define start of rains in a given period. The period is given in days in the \code{rain_day_interval} parameter.}

\item{rain_day_interval}{\code{numerical(1)} Only if \code{number_rain_days = TRUE}, the interval in days that the \code{number_rain_days} is defined in.}

\item{dry_spell}{\code{logical(1)} default \code{FALSE}. If \code{TRUE}, define start of the rains by a maximum number of dry days (\code{spell_max_dry_days}) over a given period of days (\code{spell_interval}).}

\item{spell_interval}{\code{numerical(1)} Only if \code{dry_spell = TRUE}, the interval in days that the \code{dry_spell} is defined in.}

\item{spell_max_dry_days}{\code{numerical(1)} Only if \code{dry_spell = TRUE}. This is the maximum number of dry days to define start of rains in a given period. The period is given in days in the \code{spell_interval} parameter.}

\item{dry_period}{\code{logical(1)} default \code{FALSE}. If \code{TRUE}, define start of the rains by the maximum rain and maximum dry days in a given interval. The maximum rainfall amount is given in the \code{max_rain} parameter, the maximum dry days is given in the \code{period_max_dry_days} parameter, and the interval length is given in the \code{period_interval} parameter.}

\item{period_interval}{\code{numerical(1)} Only if \code{dry_period = TRUE}, the interval in days that the \code{dry_period} is defined in.}

\item{max_rain}{\code{numerical(1)} Only if \code{dry_period = TRUE}, the maximum rainfall to occur in a given period.}

\item{period_max_dry_days}{\code{numerical(1)} Only if \code{dry_period = TRUE}. the maximum period of dry days to occur in a given period.}
}
\value{
A data.frame with the date and/or day of year for the start of the rains for each year
}
\description{
Used to present or summarise daily rainfall data in a way that corresponds to the start of the rainy season for many definitions.
}
\examples{
#TODO#
# check against R-Instat function
}
